start refactor of disk resize task
